fix: correct similarity type to REAL (float)

+++ app/src/main/java/com/fpf/smartscan/data/migrations/MIGRATION_4_5.kt
@@ -23,7 +23,7 @@ val MIGRATION_4_5 = object : Migration(4, 5) {
                 mediaId INTEGER NOT NULL,
                 conceptId INTEGER NOT NULL,
                 mediaType INTEGER NOT NULL,
-                similarity INTEGER NOT NULL,
+                similarity REAL NOT NULL,
                 PRIMARY KEY(mediaId, mediaType, conceptId),
                 FOREIGN KEY(conceptId) REFERENCES concept(id) ON DELETE CASCADE,
                 FOREIGN KEY(mediaId, mediaType) REFERENCES media_metadata(id, type) ON DELETE CASCADE
